一定時刻で現在時刻を更新するフック
